Add files via upload
Here is the original project that I had turned in that was in an inoperable condition. By removing an expense the balance becomes broken because instead of using mathematical calculations properly reimburse the balance, it instead combines the balance and the expense removed as if they both were strings and not numbers. This is just one of the many bugs on the original. I believe there are also complications removing expenses, some become complete unremovable and the expenses immediately drops to zero upon removing just one expense!