Add files via upload
Although not perfect, some fixes have been made to the budget application. This includes removing, expenses. Everything works fine until the user refreshes the page and starts removing expenses from their list of expenses. The balance then displays 'NaN' instead of the reimbursed balance. In addition to this the budget application in its current state lacks of any data validation! Overall, this version is better than the version in the Jesse branch.